refactor(ads): centralise constants and split SIMID transport into SimidRpcChannel
Reconciles the safe structural refactors from the premium ads-pro package into
the open-source ads package (behaviour unchanged; premium-only features left out):

- Add `constants.ts` with the named thresholds/limits (skip near-end, seek delta,
  non-linear default duration, poll interval, SIMID default dimensions) and replace
  the inline magic numbers in ad-dom.ts, strategies/csai.ts, and simid.ts.
  (Excluded premium's `NON_LINEAR_DISMISS_MAX_WAIT_S` clamp — that's a behavioural
  change, not a refactor.)
- Split the 751-line simid.ts into three files by extracting the module's own code:
  - `simid-protocol.ts` — SIMID 1.2 message constants + wire types.
  - `simid-rpc.ts` — `SimidRpcChannel`: postMessage transport (counter, pending
    promises, format detection, origin pinning, send/resolve/reject, inbound parsing,
    listener lifecycle).
  - `simid.ts` — `SimidSession extends SimidRpcChannel`: session handshake, media
    bridging, and lifecycle. Public API (SimidSession, SIMID_* constants,
    SimidCallbacks) is preserved via re-exports.

All 1148 tests pass; coverage 96.14%/85.86% (new files: protocol 100%, rpc 98.7%,
session 96.9%); type-check and lint clean.
